Add go build tag support to run acceptance test
