+++ .golangci.yml
@@ -63,6 +63,23 @@ run:
   # include test files or not, default is true
   tests: true
   deadline: 5m
+  build-tags:
+    - akscluster
+    - cluster
+    - clustergroup
+    - credential
+    - ekscluster
+    - gitrepository
+    - iampolicy
+    - kustomization
+    - namespace
+    - custompolicy
+    - imagepolicy
+    - networkpolicy
+    - quotapolicy
+    - securitypolicy
+    - sourcesecret
+    - workspace
 
 linters-settings:
   funlen:
