+++ internal/resources/akscluster/resource_akscluster_acc_test.go
@@ -1,3 +1,6 @@
+//go:build akscluster
+// +build akscluster
+
 /*
 Copyright 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
