+++ internal/resources/cluster/resource_cluster_test.go
@@ -1,3 +1,6 @@
+//go:build cluster
+// +build cluster
+
 /*
 Copyright © 2022 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
