+++ internal/resources/clustergroup/data_source_cluster_group_test.go
@@ -1,3 +1,6 @@
+//go:build clustergroup
+// +build clustergroup
+
 /*
 Copyright © 2021 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
