+++ internal/resources/credential/data_source_credential_test.go
@@ -1,3 +1,6 @@
+//go:build credential
+// +build credential
+
 /*
 Copyright © 2022 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
