+++ internal/resources/ekscluster/resource_ekscluster_test.go
@@ -1,3 +1,6 @@
+//go:build ekscluster
+// +build ekscluster
+
 /*
 Copyright © 2022 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
