+++ internal/resources/gitrepository/data_source_git_repository_test.go
@@ -1,3 +1,6 @@
+//go:build gitrepository
+// +build gitrepository
+
 /*
 Copyright © 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
