+++ internal/resources/gitrepository/resource_git_repository_test.go
@@ -1,3 +1,6 @@
+//go:build gitrepository
+// +build gitrepository
+
 /*
 Copyright © 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
@@ -6,17 +9,14 @@ SPDX-License-Identifier: MPL-2.0
 package gitrepository
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"os"
 	"strings"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/pkg/errors"
 
@@ -28,25 +28,6 @@ import (
 	gitrepositoryscope "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/gitrepository/scope"
 	testhelper "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/testing"
 )
-
-const (
-	gitRepositoryResource      = ResourceName
-	gitRepositoryResourceVar   = "test_git_repository"
-	gitRepositoryDataSourceVar = "test_data_source_git_repository"
-	gitRepositoryNamePrefix    = "tf-gr-test"
-)
-
-type testAcceptanceConfig struct {
-	Provider                    *schema.Provider
-	GitRepositoryResource       string
-	GitRepositoryResourceVar    string
-	GitRepositoryResourceName   string
-	GitRepositoryName           string
-	ScopeHelperResources        *commonscope.ScopeHelperResources
-	GitRepositoryDataSourceVar  string
-	GitRepositoryDataSourceName string
-	Namespace                   string
-}
 
 func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 	return &testAcceptanceConfig{
@@ -60,14 +41,6 @@ func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 		GitRepositoryDataSourceName: fmt.Sprintf("data.%s.%s", ResourceName, gitRepositoryDataSourceVar),
 		Namespace:                   "tanzu-continuousdelivery-resources",
 	}
-}
-
-func getConfigureContextFunc() func(_ context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
-	if _, found := os.LookupEnv("ENABLE_GITREPO_ENV_TEST"); !found {
-		return authctx.ProviderConfigureContextWithDefaultTransportForTesting
-	}
-
-	return authctx.ProviderConfigureContext
 }
 
 func getSetupConfig(config *authctx.TanzuContext) error {
