+++ internal/resources/iampolicy/resource_iam_policy_test.go
@@ -1,3 +1,6 @@
+//go:build iampolicy
+// +build iampolicy
+
 /*
 Copyright © 2022 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
