+++ internal/resources/namespace/data_source_namespace_test.go
@@ -1,3 +1,6 @@
+//go:build namespace
+// +build namespace
+
 /*
 Copyright © 2021 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
