+++ internal/resources/namespace/resource_namepace_test.go
@@ -1,3 +1,6 @@
+//go:build namespace
+// +build namespace
+
 /*
 Copyright © 2021 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
