+++ internal/resources/policy/kind/custom/resource/resource_custom_policy_test.go
@@ -1,3 +1,6 @@
+//go:build custompolicy
+// +build custompolicy
+
 /*
 Copyright © 2022 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
