+++ internal/resources/policy/kind/image/resource/resource_image_policy_test.go
@@ -1,3 +1,6 @@
+//go:build imagepolicy
+// +build imagepolicy
+
 /*
 Copyright © 2022 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
