+++ internal/resources/policy/kind/network/resource/resource_network_policy_test.go
@@ -1,3 +1,6 @@
+//go:build networkpolicy
+// +build networkpolicy
+
 /*
 Copyright © 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
@@ -6,17 +9,14 @@ SPDX-License-Identifier: MPL-2.0
 package networkpolicyresource
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"os"
 	"strings"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/pkg/errors"
 
@@ -31,21 +31,6 @@ import (
 	testhelper "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/testing"
 )
 
-const (
-	networkPolicyResource    = policykindNetwork.ResourceName
-	networkPolicyResourceVar = "test_network_policy"
-	networkPolicyNamePrefix  = "tf-np-test"
-)
-
-type testAcceptanceConfig struct {
-	Provider                  *schema.Provider
-	NetworkPolicyResource     string
-	NetworkPolicyResourceVar  string
-	NetworkPolicyResourceName string
-	NetworkPolicyName         string
-	ScopeHelperResources      *policy.ScopeHelperResources
-}
-
 func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 	return &testAcceptanceConfig{
 		Provider:                  initTestProvider(t),
@@ -55,15 +40,6 @@ func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 		NetworkPolicyName:         acctest.RandomWithPrefix(networkPolicyNamePrefix),
 		ScopeHelperResources:      policy.NewScopeHelperResources(),
 	}
-}
-
-// Function to set context containing different env variables based on the type of testing- Mock/Actual environment.
-func getConfigureContextFunc() func(_ context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
-	if _, found := os.LookupEnv("ENABLE_POLICY_ENV_TEST"); !found {
-		return authctx.ProviderConfigureContextWithDefaultTransportForTesting
-	}
-
-	return authctx.ProviderConfigureContext
 }
 
 // Function to set up config based on the type of testing- Mock/Actual environment.
