+++ internal/resources/policy/kind/quota/resource/resource_quota_policy_test.go
@@ -1,3 +1,6 @@
+//go:build quotapolicy
+// +build quotapolicy
+
 /*
 Copyright © 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
