+++ internal/resources/policy/kind/security/resource/resource_security_policy_test.go
@@ -1,3 +1,6 @@
+//go:build securitypolicy
+// +build securitypolicy
+
 /*
 Copyright © 2022 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
