+++ internal/resources/workspace/resource_workspace_test.go
@@ -1,3 +1,6 @@
+//go:build workspace
+// +build workspace
+
 /*
 Copyright © 2021 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
